organized directories for tracks and headings
